requirements-dev.txt
dynaconf
python-telegram-bot
cx_Oracle
pyboleto